+++ main.py
@@ -45,7 +45,8 @@ class AnotherApp(Tag.body):
 # from htag.runners import BrowserStarletteWS as Runner
 # from htag.runners import BrowserStarletteHTTP as Runner
 # from htag.runners import BrowserTornadoHTTP as Runner
-from htag.runners import ChromeApp as Runner
+# from htag.runners import ChromeApp as Runner
+from htag.runners import AndroidApp as Runner
 
 
 app=Runner(App)